requirements.txt
altgraph==0.17.4
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
packaging==25.0
pefile==2023.2.7
psutil==7.0.0
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pywin32==310
pywin32-ctypes==0.2.3
setuptools==80.9.0
Werkzeug==3.1.3
